backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
